pyproject.toml
[project]
name = "spec-trace"
version = "0.1.0"
description = "3GPP specification tracing and analysis tool"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "grandalf>=0.8",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.7",
    "langchain-google-genai>=3.1.0",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.7",
    "llama-index>=0.12.0",
    "llama-index-graph-stores-neo4j>=0.4.0",
    "llama-index-llms-openrouter>=0.4.0",
    "lxml>=6.0.2",
    "neo4j>=5.26.0",
    "openpyxl>=3.1.5",
    "python-dateutil>=2.8.0",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0",
    "requests>=2.32.5",
]
